fix: add safe integer type conversions to prevent overflow vulnerabilities
Add boundary checks for integer type conversions flagged by gosec:
- Create SafeUint64ToInt64 and SafeInt64ToUint64 helper functions
- Fix uint64 to int64 conversions in traffic cache and payment logic
- Fix int64 to uint64 conversions in traffic aggregation and stats
- Fix uint to int conversions in rate limiting middleware
- Add #nosec annotations for port range conversions (bounded by constants)

These changes prevent potential integer overflow issues while maintaining
existing business logic behavior for all practical value ranges.

+++ internal/application/telegram/admin/usecases/sendweeklysummary.go
@@ -455,6 +455,7 @@ func calculateChangePercent(current, previous int64) float64 {
 }
 
 // calculateChangePercentUint64 calculates the percentage change for uint64 values
+// Uses safe conversion to prevent integer overflow when values exceed math.MaxInt64
 func calculateChangePercentUint64(current, previous uint64) float64 {
 	if previous == 0 {
 		if current == 0 {
@@ -462,7 +463,9 @@ func calculateChangePercentUint64(current, previous uint64) float64 {
 		}
 		return 100
 	}
-	return float64(int64(current)-int64(previous)) / float64(previous) * 100
+	// Safe conversion: use float64 directly to avoid int64 overflow
+	// when current or previous exceeds math.MaxInt64
+	return (float64(current) - float64(previous)) / float64(previous) * 100
 }
 
 // formatChangeIndicator formats the change percentage with trend indicator
